Healthz endpoint library
Healthz is a set of endpoints that provides extra fuctionality for the
application to work with a a scheduler as well as to provide additional
information back to the developers of the application though profiling
systems such as pprof. Healthz attempts to provide extended
functionality for operating the application through both information as
well as direct control.

- Adds endpoint for liveliness along with functions to toggle liveliness
  status of the application using a set of provided functions
- Adds endpoint for readiness along with functions to toggle the
  readiness status of the application using a set of provided functions
- Adds endpoints to terminate the application either through the
  scheduler or by the application directly
- Adds the net/pprof endpoints underneath the the /healthz prefix
- Adds support for reporting reporter provided metrics to the developer
  through a web interface
- Adds support for tracking access to the healthz endpoints through
  DataDog metrics